investment_detector/requirements.txt
python-dotenv>=0.19.0
aiohttp>=3.8.0
pyairtable>=1.3.0
mistralai>=0.0.1
pytest>=7.0.0
pytest-asyncio>=0.18.0 